Replace the `MockStateHistoryReader` with `MockCommonState`, linter

+++ rpc/v9/subscriptions_test.go
@@ -96,7 +96,9 @@ func (fs *fakeSyncer) PendingState() (core.CommonStateReader, func() error, erro
 	return nil, nil, nil
 }
 
-func (fs *fakeSyncer) PendingStateBeforeIndex(index int) (core.CommonStateReader, func() error, error) {
+func (fs *fakeSyncer) PendingStateBeforeIndex(
+	index uint64,
+) (core.CommonStateReader, func() error, error) {
 	return nil, nil, nil
 }
